move batches tests to own file
